Remove reverse substring behavior and change args to longs

+++ include/ilc/string.h
@@ -12,7 +12,7 @@ String *create_string(const char *chars, size_t len);
 String *string_copy(const String *str);
 String *string_reverse(const String *str);
 void free_string(String *str);
-String *substring(const String *str, int start, int end);
+String *substring(const String *str, long start, long end);
 char *string_to_c_string(const String *str);
 int string_equal(const String *str1, const String *str2);
 int string_compare(const String *str1, const String *str2);
